# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Board API",
    "version": "1.0.0",
    "description": "\n# 채용 플랫폼 API 문서\n\n## 인증 방식\n- Bearer Token 인증 사용\n- /auth/login에서 발급받은 access_token을 Authorization 헤더에 포함\n- 예: Authorization: Bearer <access_token>\n\n## 공통 에러 코드\n- 400: 잘못된 요청\n- 401: 인증 실패\n- 403: 권한 없음\n- 404: 리소스 없음\n- 409: 충돌 (중복 등)\n- 500: 서버 에러\n\n## 응답 형식\n성공 응답:\n```json\n{\n    \"status\": \"success\",\n    \"data\": { ... }\n}\n```\n\n에러 응답:\n```json\n{\n    \"status\": \"error\",\n    \"message\": \"에러 메시지\"\n}\n```\n            "
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "채용공고 목록 조회",
        "description": "채용공고 목록을 검색, 필터링, 정렬하여 조회합니다.",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "검색어 (제목, 내용)"
          },
          {
            "in": "query",
            "name": "location_id",
            "schema": {
              "type": "integer"
            },
            "description": "지역 ID"
          },
          {
            "in": "query",
            "name": "categories",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "description": "직무 카테고리 ID 목록",
            "style": "form",
            "explode": false
          },
          {
            "in": "query",
            "name": "tech_stacks",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "description": "기술 스택 ID 목록",
            "style": "form",
            "explode": false
          },
          {
            "in": "query",
            "name": "sort_by",
            "schema": {
              "type": "string",
              "enum": [
                "latest",
                "views",
                "deadline"
              ]
            },
            "description": "정렬 기준"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "페이지 번호"
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "페이지당 항목 수"
          }
        ],
        "responses": {
          "200": {
            "description": "채용공고 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "postings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/JobPosting"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "per_page": {
                          "type": "integer"
                        },
                        "total_pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "채용공고 등록",
        "description": "새로운 채용공고를 등록합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPostingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "채용공고 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "posting_id": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{posting_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "채용공고 상세 조회",
        "description": "특정 채용공고의 상세 정보를 조회합니다.",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "채용공고 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "채용공고 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/JobPosting"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "채용공고 수정",
        "description": "특정 채용공고의 정보를 수정합니다. 회사 권한이 필요합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "채용공고 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "수정된 채용공고 제목"
                  },
                  "job_description": {
                    "type": "string",
                    "example": "수정된 채용공고 내용"
                  },
                  "experience_level": {
                    "type": "string",
                    "example": "경력 3년 이상"
                  },
                  "education_level": {
                    "type": "string",
                    "example": "학사 이상"
                  },
                  "employment_type": {
                    "type": "string",
                    "example": "정규직"
                  },
                  "salary_info": {
                    "type": "string",
                    "example": "4,500-5,500만원"
                  },
                  "location_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "deadline_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-31"
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      1,
                      3
                    ]
                  },
                  "tech_stacks": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      2,
                      4,
                      6
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "채용공고 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Job posting updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 또는 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Posting not found or unauthorized"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "채용공고 삭제",
        "description": "특정 채용공고를 삭제(비활성화)합니다. 회사 권한이 필요합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "채용공고 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "채용공고 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Job posting deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 또는 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Posting not found or unauthorized"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Error message here"
          }
        }
      },
      "JobPosting": {
        "type": "object",
        "properties": {
          "posting_id": {
            "type": "integer"
          },
          "company_id": {
            "type": "integer"
          },
          "company_name": {
            "type": "string"
          },
          "company_description": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "job_description": {
            "type": "string"
          },
          "experience_level": {
            "type": "string"
          },
          "education_level": {
            "type": "string"
          },
          "employment_type": {
            "type": "string"
          },
          "salary_info": {
            "type": "string"
          },
          "location_id": {
            "type": "integer"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "deadline_date": {
            "type": "string",
            "format": "date"
          },
          "view_count": {
            "type": "integer"
          },
          "categories": {
            "type": "string"
          },
          "tech_stacks": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JobPostingInput": {
        "type": "object",
        "required": [
          "title",
          "job_description"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "백엔드 개발자 모집"
          },
          "job_description": {
            "type": "string",
            "example": "Python/Django 백엔드 개발자를 모집합니다."
          },
          "experience_level": {
            "type": "string",
            "enum": ["신입", "경력", "무관"],
            "example": "신입"
          },
          "education_level": {
            "type": "string",
            "enum": ["고졸", "초대졸", "학사", "석사", "박사"],
            "example": "학사"
          },
          "employment_type": {
            "type": "string",
            "enum": ["정규직", "계약직", "인턴"],
            "example": "정규직"
          },
          "salary_info": {
            "type": "string",
            "example": "4,000-5,000만원"
          },
          "location": {
            "type": "object",
            "properties": {
              "region": {
                "type": "string",
                "enum": ["수도권", "광역시", "도", "해외"],
                "example": "수도권"
              },
              "sub_region": {
                "type": "string",
                "enum": [
                  "서울", "경기", "인천",
                  "부산", "대구", "광주", "대전", "울산", "세종",
                  "강원", "경남", "경북", "전남", "전북", "충남", "충북", "제주",
                  "아시아", "북미", "유럽", "오세아니아"
                ],
                "example": "서울"
              }
            }
          },
          "deadline_date": {
            "type": "string",
            "format": "date",
            "example": "2024-12-31"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string",
                  "enum": ["개발", "기획", "디자인"],
                  "example": "개발"
                },
                "sub_category": {
                  "type": "string",
                  "enum": [
                    "백엔드", "프론트엔드", "풀스택", "모바일", "데이터", "DevOps",
                    "서비스기획", "프로젝트관리", "전략기획",
                    "UX/UI", "웹디자인", "그래픽"
                  ],
                  "example": "백엔드"
                }
              }
            },
            "example": [
              {
                "category": "개발",
                "sub_category": "백엔드"
              },
              {
                "category": "개발",
                "sub_category": "풀스택"
              }
            ]
          },
          "tech_stacks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Python", "Django", "MySQL"]
          }
        }
      },
...
    },
...
  },
...
}